Pass Data Analysis tab into Fit Plot presenter
